Fixed the comment functionality.
